Refactor `InferenceModel`
